Removed python related code from main sprokit library.
Having the python code in the main library caused a unwanted
dependency on python libraries. The python related code is moved
to the python bindings.

Moved python factories to the bindings directory.
