+++ sprokit/src/sprokit/pipeline/scheduler_factory.cxx
@@ -139,35 +139,4 @@ is_scheduler_module_loaded( kwiver::vital::plugin_loader& vpl,
   return vpl.is_module_loaded( mod );
 }
 
-// ------------------------------------------------------------------
-#ifdef SPROKIT_ENABLE_PYTHON
-
-python_scheduler_factory::
-python_scheduler_factory( const std::string& type,
-                          const std::string& itype,
-                          py_scheduler_factory_func_t factory )
-  : scheduler_factory( type, itype )
-  , m_factory( factory )
-{
-  this->add_attribute( CONCRETE_TYPE, type)
-    .add_attribute( PLUGIN_FACTORY_TYPE, typeid(* this ).name() )
-    .add_attribute( PLUGIN_CATEGORY, "scheduler" );
-}
-
-
-// ----------------------------------------------------------------------------
-scheduler_t
-python_scheduler_factory::
-create_object(sprokit::pipeline_t const& pipe, kwiver::vital::config_block_sptr const& config)
-{
-  // Call sprokit factory function.
-  pybind11::object obj = m_factory(pipe, config);
-  obj.inc_ref();
-  sprokit::scheduler_t schd_ptr = obj.cast<sprokit::scheduler_t>();
-  return schd_ptr;
-}
-
-#endif
-
-
 } // end namespace
