+++ sprokit/src/sprokit/pipeline/scheduler_factory.h
@@ -47,10 +47,6 @@
 #include <functional>
 #include <memory>
 
-#ifdef SPROKIT_ENABLE_PYTHON
-  #include <pybind11/pybind11.h>
-#endif
-
 namespace sprokit {
 
 // returns: scheduler_t - shared_ptr<scheduler>
@@ -191,32 +187,6 @@ bool is_scheduler_module_loaded( kwiver::vital::plugin_loader& vpl,
   add_factory( new sprokit::cpp_scheduler_factory( typeid( type ).name(), \
                                                    typeid( sprokit::scheduler ).name(), \
                                                    sprokit::create_new_scheduler< type > ) )
-
-#ifdef SPROKIT_ENABLE_PYTHON
-
-typedef std::function< pybind11::object( sprokit::pipeline_t const& pipe,
-                               kwiver::vital::config_block_sptr const& config ) > py_scheduler_factory_func_t;
-
-class SPROKIT_PIPELINE_EXPORT python_scheduler_factory
-: public scheduler_factory
-{
-
-  public:
-
-  python_scheduler_factory( const std::string& type,
-                            const std::string& itype,
-                            py_scheduler_factory_func_t factory );
-
-  virtual ~python_scheduler_factory() = default;
-
-  virtual scheduler_t create_object(sprokit::pipeline_t const& pipe,
-                                    kwiver::vital::config_block_sptr const& config);
-
-private:
-  py_scheduler_factory_func_t m_factory;
-};
-
-#endif
 
 } // end namespace
